Fix lift/PID
Fix lift::getCurrentPos() using encoder values instead of 0-1 as
advertised in header.

Protect leftPos/rightPos global vars using mutexes.

Raise liftIncrement constant to be noticeable.

+++ src/pid.cpp
@@ -47,7 +47,7 @@ int PID::update(int value, int deltaTime)
 
     // power clamped to the interval [-127, 127]
     int power = round(127 * tanh(p));
-    printf("value: %d, p: %.2f, power: %d\n", value, (double) p, power);
+    printf("value: %d, p: %.2f, power: %d\n", value, p, power);
     return power;
 }
